Organização da estrutura de pastas , Inicio de criação do footer
