+++ src/App.jsx
@@ -1,16 +1,16 @@
 import './App.css';
 
 import { Routes, Route } from 'react-router-dom';
-import { AdminPageContextProvider } from './assets/Contexts/AdminContext';
-import NavAndSid_Bar from './NavAndSid_Bar';
-
-import Home from './assets/Components/Home';
-import Admin from './assets/Components/Admin';
-import Products from './assets/Components/Products';
-import Categories from './assets/Components/Categories';
-import Carrinho from './assets/Components/Carrinho';
-import Sobre from './assets/Components/Sobre';
-import Errorpg from './assets/Components/Errorpg';
+import { AdminPageContextProvider } from './Contexts/AdminContext';
+import NavAndSid_Bar from './Components/Header/NavAndSid_Bar';
+import Footer from './Components/Footer/Footer';
+import Home from './pages/home_pg/Home';
+import Admin from './pages/admin_pg/Admin';
+import Products from './pages/produtct_pg/Products';
+import Categories from './pages/categories_pg/Categories';
+import Carrinho from './pages/carrinho_pg/Carrinho';
+import Sobre from './pages/sobre_pg/Sobre';
+import Errorpg from './pages/error_pg/Errorpg';
 
 
 function App() {
@@ -20,8 +20,8 @@ function App() {
   <NavAndSid_Bar/>
   <AdminPageContextProvider>
   <Routes>
+    <Route path="/" element={<Home />}/>
     <Route path='/pg_admin' element={<Admin />}/>
-    <Route path="/pg_home" element={<Home />}/>
     <Route path='/pg_products' element={<Products />}/>
     <Route path='/pg_categories' element={<Categories />}/>
     <Route path='/pg_carrinho' element={<Carrinho />}/>
@@ -29,6 +29,7 @@ function App() {
     <Route path='/pg_404' element={<Errorpg />}/>
   </Routes>
   </AdminPageContextProvider>
+  <Footer/>
   </>
   )
 }
