+++ src/Components/Footer/Footer.css
@@ -1,0 +1,8 @@
+.CorNav {
+  background-color: #FFDE59;
+  box-shadow: 0 8px 32px 0 rgba( 31, 38, 135, 0.37 );
+  backdrop-filter: blur( 9px );
+  -webkit-backdrop-filter: blur( 9px );
+  border-radius: 0px 0px 10px 10px;
+  border: 1px solid rgba( 255, 255, 255, 0.18 );
+}
