+++ src/Components/Footer/Footer.jsx
@@ -1,0 +1,15 @@
+import Container from "react-bootstrap/Container";
+import "./Footer.css";
+
+export default function Footer() {
+  return (
+    <footer className="CorNav w-100">
+    
+        <div> teste</div>
+      
+      
+    </footer>
+  );
+}
+
+
