--- src/assets/imgs/Import_Imgs.js
@@ -1,6 +1,0 @@
-export const imagemItem_1 = './wallpaperflare.com_wallpaper.jpg'
-export const imagemItem_2 = './wallpaperflare.com_wallpaper(1).jpg'
-export const imagemItem_3 = './wallpaperflare.com_wallpaper(3).jpg'
-export const imagemItem_4 = './wallpaperflare.com_wallpaper(4).jpg'
-
-
